prepare for contrib/float16
