+++ python/paddle/fluid/tests/demo/float16/run_float16_demo.sh
@@ -8,16 +8,16 @@ DEMO_PATH=/paddle/python/paddle/fluid/tests/demo/float16
 export CUDA_VISIBLE_DEVICES=0
 
 # Build the PaddlePaddle Fluid wheel package and install it.
-mkdir -p $BUILD_PATH && cd $BUILD_PATH
-cmake .. -DWITH_AVX=OFF \
-         -DWITH_MKL=OFF \
-         -DWITH_GPU=ON \
-         -DWITH_TESTING=ON \
-         -DWITH_TIMER=ON \
-         -DWITH_PROFILER=ON \
-         -DWITH_FLUID_ONLY=ON
-make -j `nproc`
-pip install -U "$WHEEL_PATH/$(ls $WHEEL_PATH)"
+#mkdir -p $BUILD_PATH && cd $BUILD_PATH
+#cmake .. -DWITH_AVX=OFF \
+#         -DWITH_MKL=OFF \
+#         -DWITH_GPU=ON \
+#         -DWITH_TESTING=ON \
+#         -DWITH_TIMER=ON \
+#         -DWITH_PROFILER=ON \
+#         -DWITH_FLUID_ONLY=ON
+#make -j `nproc`
+#pip install -U "$WHEEL_PATH/$(ls $WHEEL_PATH)"
 
 cd $DEMO_PATH
